AIQA/requirements.txt
# Core
numpy==1.26.4
pillow==10.4.0
rapidfuzz==3.9.7
PyYAML==6.0.2
opencv-python==4.10.0.84
selenium==4.23.1
streamlit==1.37.1

# Ultralytics (YOLOv12-capable; supports C3k2 blocks)
ultralytics>=8.3.78,<9

# Torch CPU (py311). If install fails, install it separately:
#   python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# OCR backends — pick at least one:
pytesseract
# OR:
# paddlepaddle==2.6.1
# paddleocr==2.7.0.3
# OR:
# easyocr
